+++ docs/history.rst
@@ -24,6 +24,7 @@ Version 1.7.0 - Unreleased
 * Remove Sphinx from requirement in development.cfg, close #76;
 * Fixed cookieconsent template to use only simple quote ``'`` to surround variable strings, close #75;
 * Upgraded to ``crispy-forms-foundation==0.5.5``;
+* Removed unused ``pages/3_cols.html`` template and updated CMS settings to enable page template ``pages/free.html``;
 
 Version 1.6.0 - 2016/12/17
 --------------------------